Add CurveInterval, change time long->double halfways through

+++ src/jcurl/core/RockSetInterpolator.java
@@ -67,15 +67,15 @@ public class RockSetInterpolator implements Source, TargetDiscrete {
             }
     }
 
-    public long getMaxT() {
+    public double getMaxT() {
         return ip[0].getMaxT();
     }
 
-    public long getMinT() {
+    public double getMinT() {
         return ip[0].getMinT();
     }
 
-    public RockSet getPos(final long t, final RockSet rocks) {
+    public RockSet getPos(final double t, final RockSet rocks) {
         if (t < ip[0].getMinT() || t > ip[0].getMaxT())
             return null;
         final RockSet dat = rocks != null ? rocks : new RockSet();
@@ -84,7 +84,7 @@ public class RockSetInterpolator implements Source, TargetDiscrete {
         return dat;
     }
 
-    public RockSet getSpeed(final long t, final RockSet rocks) {
+    public RockSet getSpeed(final double t, final RockSet rocks) {
         if (t < ip[0].getMinT() || t > ip[0].getMaxT())
             return null;
         final RockSet dat = rocks != null ? rocks : new RockSet();
@@ -108,18 +108,18 @@ public class RockSetInterpolator implements Source, TargetDiscrete {
         return true;
     }
 
-    public void reset(long startTime, RockSet startPos, RockSet startSpeed,
+    public void reset(double startTime, RockSet startPos, RockSet startSpeed,
             RockSetProps props) {
         for (int i = RockSet.ROCKS_PER_SET - 1; i >= 0; i--)
             this.ip[i].reset();
         setPos(startTime, startPos);
     }
 
-    public void setPos(final long t, final RockSet rocks) {
+    public void setPos(final double t, final RockSet rocks) {
         setPos(t, rocks, 0);
     }
 
-    public void setPos(final long t, final RockSet rocks,
+    public void setPos(final double t, final RockSet rocks,
             final int discontinuous) {
         for (int i = RockSet.ROCKS_PER_SET - 1; i >= 0; i--) {
             ip[i].add(t, rocks.getRock(i), 0 != ((1 << i) | discontinuous));
